moved function_traits to detail directory
the functions traits (while quite nicely done)
are an implementation detail of the metaprogramming in flexcore.
